form changed

+++ database/migrations/2018_09_27_123126_create_depreciation_types_table.php
@@ -17,7 +17,7 @@ class CreateDepreciationTypesTable extends Migration
             $table->increments('dp_type_id');
             $table->integer('current_dp');
             $table->integer('old_value');
-            $table->integer('dp_ratio');
+            $table->decimal('dp_ratio');
             $table->timestamps();
         });
     }
